requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
email-validator==2.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
SQLAlchemy==2.0.21
WTForms==3.0.1
python-dotenv==1.0.0
Flask-Migrate==4.0.5
